# *: Add trySeekUsingNext optimization to SeekGE