--- a/ingest.go
+++ b/ingest.go
@@ -333,7 +333,7 @@ func overlapWithIterator(
 	//    means boundary < L and hence is similar to 1).
 	// 4) boundary == L and L is sentinel,
 	//    we'll always overlap since for any values of i,j ranges [i, k) and [j, k) always overlap.
-	key, _ := iter.SeekGE(meta.Smallest.UserKey)
+	key, _ := iter.SeekGE(meta.Smallest.UserKey, false /* trySeekUsingNext */)
 	if key != nil {
 		c := sstableKeyCompare(cmp, *key, meta.Largest)
 		if c <= 0 {