--- a/testdata/iterator
+++ b/testdata/iterator
@@ -198,6 +198,16 @@ b:c
 .
 stats: (interface (dir, seek, step): (fwd, 3, 0), (rev, 0, 0)), (internal (dir, seek, step): (fwd, 3, 0), (rev, 0, 0))
 
+iter seq=3
+seek-ge a
+seek-ge b
+seek-ge c
+----
+a:b
+b:c
+.
+stats: (interface (dir, seek, step): (fwd, 3, 0), (rev, 0, 0)), (internal (dir, seek, step): (fwd, 3, 0), (rev, 0, 0))
+
 define
 a.SET.1:a
 b.SET.2:b