This change takes the trySeekUsingNext optimization that
exists in SeekPrefixGE and applies that to SeekGE as well. It
is expected to help out a lot in cockroachdb/cockroach#66410
as the CheckSSTCollisions call in Cockroach repeatedly
calls SeekGE on the engine iterator in bad cases of
ingested sst <> engine overlap.


Also adds a small optimization to not reload a block in blockIter
if we landed on the same data block after a Seek.