@itsbilal 2021-12-15
Running CheckSSTConflicts with and without this optimization against a pre-https://github.com/cockroachdb/cockroach/pull/73514 version of Cockroach gets me:

```
name                                                                      old time/op  new time/op  delta
CheckSSTConflicts/keys=1000/versions=8/sstKeys=1000/overlap=false-24      72.6µs ±11%  69.2µs ± 1%     ~     (p=0.690 n=5+5)
CheckSSTConflicts/keys=1000/versions=8/sstKeys=1000/overlap=true-24       12.2ms ± 1%   1.7ms ± 0%  -86.07%  (p=0.008 n=5+5)
CheckSSTConflicts/keys=1000/versions=8/sstKeys=10000/overlap=false-24     69.8µs ± 2%  70.3µs ± 0%     ~     (p=0.151 n=5+5)
CheckSSTConflicts/keys=1000/versions=8/sstKeys=10000/overlap=true-24      13.3ms ± 3%   2.9ms ± 1%  -78.51%  (p=0.008 n=5+5)
CheckSSTConflicts/keys=1000/versions=64/sstKeys=1000/overlap=false-24     75.8µs ± 3%  67.1µs ± 1%  -11.50%  (p=0.008 n=5+5)
CheckSSTConflicts/keys=1000/versions=64/sstKeys=1000/overlap=true-24      13.0ms ± 1%   2.0ms ± 3%  -84.30%  (p=0.008 n=5+5)
CheckSSTConflicts/keys=1000/versions=64/sstKeys=10000/overlap=false-24    69.8µs ±11%  67.2µs ± 1%     ~     (p=0.690 n=5+5)
CheckSSTConflicts/keys=1000/versions=64/sstKeys=10000/overlap=true-24     14.8ms ± 9%   3.8ms ± 8%  -74.21%  (p=0.008 n=5+5)
CheckSSTConflicts/keys=10000/versions=8/sstKeys=1000/overlap=false-24     66.1µs ± 2%  66.0µs ± 1%     ~     (p=0.841 n=5+5)
CheckSSTConflicts/keys=10000/versions=8/sstKeys=1000/overlap=true-24      14.2ms ± 9%   2.1ms ± 1%  -85.51%  (p=0.016 n=5+4)
CheckSSTConflicts/keys=10000/versions=8/sstKeys=10000/overlap=false-24    72.3µs ±10%  66.6µs ± 1%     ~     (p=0.690 n=5+5)
CheckSSTConflicts/keys=10000/versions=8/sstKeys=10000/overlap=true-24      122ms ± 2%    19ms ± 1%  -84.39%  (p=0.008 n=5+5)
CheckSSTConflicts/keys=10000/versions=64/sstKeys=1000/overlap=false-24    69.0µs ± 9%  64.4µs ± 1%     ~     (p=0.690 n=5+5)
CheckSSTConflicts/keys=10000/versions=64/sstKeys=1000/overlap=true-24     14.0ms ± 1%   3.0ms ±11%  -78.55%  (p=0.016 n=4+5)
CheckSSTConflicts/keys=10000/versions=64/sstKeys=10000/overlap=false-24   69.4µs ± 9%  65.1µs ± 1%     ~     (p=0.690 n=5+5)
CheckSSTConflicts/keys=10000/versions=64/sstKeys=10000/overlap=true-24     140ms ± 5%    33ms ± 3%  -76.56%  (p=0.008 n=5+5)
CheckSSTConflicts/keys=100000/versions=8/sstKeys=1000/overlap=false-24    69.2µs ±10%  64.5µs ± 2%     ~     (p=0.310 n=5+5)
CheckSSTConflicts/keys=100000/versions=8/sstKeys=1000/overlap=true-24     15.3ms ±11%   2.7ms ±12%  -82.09%  (p=0.008 n=5+5)
CheckSSTConflicts/keys=100000/versions=8/sstKeys=10000/overlap=false-24   69.7µs ±12%  64.8µs ± 1%     ~     (p=0.548 n=5+5)
CheckSSTConflicts/keys=100000/versions=8/sstKeys=10000/overlap=true-24     148ms ± 6%    36ms ± 2%  -75.34%  (p=0.008 n=5+5)
CheckSSTConflicts/keys=100000/versions=64/sstKeys=1000/overlap=false-24   67.1µs ±10%  64.5µs ± 5%     ~     (p=1.000 n=5+5)
CheckSSTConflicts/keys=100000/versions=64/sstKeys=1000/overlap=true-24    14.4ms ± 2%   4.9ms ± 8%  -66.26%  (p=0.016 n=4+5)
CheckSSTConflicts/keys=100000/versions=64/sstKeys=10000/overlap=false-24  68.9µs ±21%  64.5µs ± 3%     ~     (p=1.000 n=5+5)
CheckSSTConflicts/keys=100000/versions=64/sstKeys=10000/overlap=true-24    204ms ±14%    61ms ±12%  -70.12%  (p=0.008 n=5+5)
```